# docs/swagger.yaml
definitions:
...
  model.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
...
  model.Game:
    properties:
      blackPlayerId:
        type: integer
      createdAt:
        type: string
      creatorId:
        type: integer
      endedAt:
        type: string
      id:
        type: integer
      inProgress:
        type: boolean
      lastMovePlayedAt:
        type: string
      name:
        type: string
      public:
        type: boolean
      startedAt:
        type: string
      tiles:
        type: string
      turnDurationSeconds:
        type: integer
      whitePlayerId:
        type: integer
      winnerId:
        type: integer
    type: object
  model.GameCreate:
    properties:
      isWhite:
        type: boolean
      name:
        type: string
      password:
        type: string
      turnDurationSeconds:
        type: integer
    type: object
  model.GameJoin:
    properties:
      password:
        type: string
    type: object
  model.GameListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/model.Game'
        type: array
      resultCount:
        type: integer
      totalCount:
        type: integer
    type: object
  model.GameMakeMove:
    properties:
      move:
        type: string
    type: object
  model.GenericResponse:
    properties:
      data:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: lukamatosevic5@gmail.com
    name: Luka Matošević
    url: https://lukamatosevic.com
  license:
    name: MIT 2023
    url: https://www.mit.edu/~amini/LICENSE.md
paths:
...
  /v1/games:
    get:
      description: Query and list games
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Size
        in: query
        name: size
        type: integer
      - description: Sort
        in: query
        name: sort
        type: string
      - description: Filter
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.GameListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Query and list games
      tags:
      - games
  /v1/games/{id}:
    get:
      description: Find one game
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.Game'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Find one game
      tags:
      - games
  /v1/games/{id}/join:
    post:
      consumes:
      - application/json
      description: Join existing game
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: integer
      - description: Join game
        in: body
        name: game
        schema:
          $ref: '#/definitions/model.GameJoin'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Join existing game
      tags:
      - games
  /v1/games/{id}/move:
    post:
      consumes:
      - application/json
      description: Make game move
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: integer
      - description: Game move
        in: body
        name: game
        required: true
        schema:
          $ref: '#/definitions/model.GameMakeMove'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Make game move
      tags:
      - games
  /v1/games/{id}/moves:
    get:
      description: Query and list game moves
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Size
        in: query
        name: size
        type: integer
      - description: Sort
        in: query
        name: sort
        type: string
      - description: Filter
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.GameListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Query and list game moves
      tags:
      - games
  /v1/games/{id}/quit:
    post:
      consumes:
      - application/json
      description: Quit joined game
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Quit joined game
      tags:
      - games
  /v1/games/create:
    post:
      consumes:
      - application/json
      description: Create new game
      parameters:
      - description: Create game
        in: body
        name: game
        required: true
        schema:
          $ref: '#/definitions/model.GameCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/model.Game'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create new game
      tags:
      - games
...
securityDefinitions:
  ApiKeyAuth:
    description: The access token obtained from /login endpoint, required for accessing
      protected routes
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"